+++ libc/src/stdio/vsnprintf.cpp
@@ -41,7 +41,7 @@ LLVM_LIBC_FUNCTION(int, vsnprintf,
   if (buffsz > 0) // if the buffsz is 0 the buffer may be a null pointer.
     wb.buff[wb.buff_cur] = '\0';
 
-  if (ret_val.value() > cpp::numeric_limits<int>::max()) {
+  if (ret_val.value() > static_cast<size_t>(cpp::numeric_limits<int>::max())) {
     libc_errno =
         printf_core::internal_error_to_errno(-printf_core::OVERFLOW_ERROR);
     return -1;
